fix remaining circular dependencies
