+++ package.json
@@ -8,11 +8,14 @@
     "analyze-circular-dependencies-fields": "madge test/fields/config.ts --circular --warning",
     "analyze-circular-dependencies-lexical": "madge packages/richtext-lexical/src/index.ts --circular --warning",
     "analyze-circular-dependencies-lexical-client": "madge packages/richtext-lexical/src/exports/client/index.ts --circular --warning",
+    "analyze-circular-dependencies-lexical-migrate": "madge packages/richtext-lexical/src/exports/server/migrate.ts --circular --warning",
     "analyze-circular-dependencies-mongodb": "madge packages/db-mongodb/src/index.ts --circular --warning",
     "analyze-circular-dependencies-postgres": "madge packages/db-postgres/src/index.ts --circular --warning",
     "analyze-circular-dependencies-drizzle": "madge packages/drizzle/src/index.ts --circular --warning",
     "analyze-circular-dependencies-drizzle-postgres": "madge packages/drizzle/src/exports/postgres.ts --circular --warning",
     "analyze-circular-dependencies-graphql": "madge packages/grapqhl/src/index.ts --circular --warning",
+    "analyze-circular-dependencies-fieldSchemasToFormState": "madge packages/ui/src/forms/fieldSchemasToFormState/index.tsx --circular --warning",
+    "analyze-circular-dependencies-buildFieldSchemaMap": "madge packages/ui/src/utilities/buildFieldSchemaMap/index.ts --circular --warning",
     "bf": "pnpm run build:force",
     "build": "pnpm run build:core",
     "build:all": "turbo build",
